Updated documentation, unified build, and CircleCI integration
