+++ .circleci/config.yml
@@ -1,0 +1,120 @@
+# Javascript Node CircleCI 2.0 configuration file
+#
+# Check https://circleci.com/docs/2.0/language-javascript/ for more details
+#
+version: 2
+jobs:
+  test:
+    docker:
+      # specify the version you desire here
+      - image: circleci/node:12-browsers
+
+      # Specify service dependencies here if necessary
+      # CircleCI maintains a library of pre-built images
+      # documented at https://circleci.com/docs/2.0/circleci-images/
+      # - image: circleci/mongo:3.4.4
+
+    working_directory: ~/repo
+
+    steps:
+      - checkout
+
+      # Download and cache dependencies
+      - restore_cache:
+          keys:
+            - v1-dependencies-{{ checksum "package.json" }}
+            # fallback to using the latest cache if no exact match is found
+            - v1-dependencies-
+
+      # Test and build the main repo
+      - run:
+          working_directory: ~/repo
+          command: |
+            npm install
+            CHROME_BIN=/usr/bin/google-chrome
+            npm test
+
+  build:
+    docker:
+      # specify the version you desire here
+      - image: circleci/node:12-browsers
+
+      # Specify service dependencies here if necessary
+      # CircleCI maintains a library of pre-built images
+      # documented at https://circleci.com/docs/2.0/circleci-images/
+      # - image: circleci/mongo:3.4.4
+
+    working_directory: ~/repo
+
+    steps:
+      - checkout
+
+      # Download and cache dependencies
+      - restore_cache:
+          keys:
+            - v1-dependencies-{{ checksum "package.json" }}
+            # fallback to using the latest cache if no exact match is found
+            - v1-dependencies-
+
+      # Test and build the main repo
+      - run:
+          working_directory: ~/repo
+          command: |
+            npm install
+            npm run build
+
+  prerelease:
+    docker:
+      # specify the version you desire here
+      - image: circleci/node:12-browsers
+
+      # Specify service dependencies here if necessary
+      # CircleCI maintains a library of pre-built images
+      # documented at https://circleci.com/docs/2.0/circleci-images/
+      # - image: circleci/mongo:3.4.4
+
+    working_directory: ~/repo
+
+    steps:
+      - checkout
+
+      # Download and cache dependencies
+      - restore_cache:
+          keys:
+            - v1-dependencies-{{ checksum "package.json" }}
+            # fallback to using the latest cache if no exact match is found
+            - v1-dependencies-
+
+      # Test and build the main repo
+      - run:
+          working_directory: ~/repo
+          command: |
+            npm install
+            npm run build
+
+workflows:
+  version: 2
+  dev:
+    jobs:
+      - test:
+          filters:
+            branches:
+              ignore: master
+      - build:
+          requires:
+            - test
+          filters:
+            branches:
+              ignore: master
+  prerelease:
+    jobs:
+      - test:
+          filters:
+            branches:
+              only: master
+      - prerelease:
+          requires:
+            - test
+          filters:
+            branches:
+              only: master
