+++ package.json
@@ -4,9 +4,12 @@
   "description": "Utility for maintaining state in an application",
   "main": "index.js",
   "scripts": {
-    "pretest": "npx webpack-serve ./test/webpack.config.js --open --port=8957",
-    "test": "echo \"Open up the test.html file in your browser.\"",
-    "build": "npx webpack ./src/application_state.js --output-filename application_state.js"
+    "dev": "npx webpack-serve ./test/webpack.config.js --open --port=8957",
+    "test": "npx karma start --single-run --browsers ChromeHeadless karma.conf.js",
+    "build:main": "npx webpack ./src/application_state.js --output-filename application_state.js",
+    "build:indexeddb": "npx webpack ./plugins/indexeddb/index.js --output-filename application_state-plugin-indexeddb.js",
+    "build:localstorage": "npx webpack ./plugins/localstorage/index.js --output-filename application_state-plugin-localstorage.js",
+    "build": "npm run build:main && npm run build:indexeddb && npm run build:localstorage"
   },
   "repository": {
     "type": "git",
@@ -29,8 +32,13 @@
     "@babel/core": "^7.5.5",
     "@babel/preset-env": "^7.5.5",
     "babel-loader": "^8.0.6",
-    "chai": "^4.1.2",
+    "chai": "^4.2.0",
     "jquery": "^3.4.1",
+    "karma": "^4.2.0",
+    "karma-chai": "^0.1.0",
+    "karma-chrome-launcher": "^3.0.0",
+    "karma-mocha": "^1.3.0",
+    "karma-webpack": "^4.0.2",
     "mocha": "^5.2.0",
     "webpack": "^4.16.4",
     "webpack-cli": "^3.1.0",
@@ -38,6 +46,7 @@
   },
   "browserslist": "> 0.25%, not dead",
   "dependencies": {
-    "core-js": "^2.6.9"
+    "core-js": "^2.6.9",
+    "dexie": "^2.0.4"
   }
 }
