--- plugins/indexeddb/.browserlistrc
@@ -1,3 +1,0 @@
-> 1%
-not dead
-ie 11
