--- plugins/indexeddb/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "applicationstate-plugin-indexeddb",
-  "version": "0.0.2",
-  "description": "IndexedDB plugin for application state loading and persisting",
-  "main": "index.js",
-  "scripts": {
-    "pretest": "node_modules/.bin/webpack-serve ./test/webpack.config.js --open --port=8958",
-    "test": "echo \"Open up the test.html file in your browser.\"",
-    "build": "node_modules/.bin/webpack ./index.js --output-filename application_state.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/claytongulick/applicationstate-plugin-indexeddb.git"
-  },
-  "author": "Clayton Gulick",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/claytongulick/applicationstate-plugin-indexeddb/issues"
-  },
-  "homepage": "https://github.com/claytongulick/applicationstate-plugin-indexeddb#readme",
-  "dependencies": {
-    "dexie": "^2.0.4"
-  },
-  "devDependencies": {
-    "chai": "^4.1.2",
-    "core-js": "^2.6.9",
-    "jquery": "^3.3.1",
-    "mocha": "^5.2.0",
-    "webpack": "^4.16.4",
-    "webpack-cli": "^3.1.0",
-    "webpack-serve": "^2.0.2"
-  }
-}
