--- plugins/indexeddb/webpack.config.js
@@ -1,16 +1,0 @@
-const webpack = require('webpack');
-const path = require('path');
-
-const core_config = {
-  mode: 'production',
-  devtool: 'none',
-  entry: {
-    bundle: './index.js',
-  },
-  output: {
-    path: path.normalize(__dirname + '/../../dist/'),
-    filename: 'applicationstate-plugin-indexeddb.js'
-  }
-}
-
-module.exports = core_config;
