+++ test/arrays.spec.js
@@ -1,4 +1,6 @@
-const expect = chai.expect;
+import ApplicationState from "../src/application_state";
+
+if (!expect) var expect = chai.expect;
 
 describe('Working with Arrays', function () {
     it('should get and set a simple array', function () {
