+++ test/basic.spec.js
@@ -1,4 +1,6 @@
-const expect = chai.expect;
+import ApplicationState from "../src/application_state";
+
+if (!expect) var expect = chai.expect;
 
 describe('Basic', function () {
     it('should set and get numeric values', function () {
