+++ test/flatten.spec.js
@@ -1,6 +1,6 @@
 import ApplicationState from "../src/application_state";
 
-const expect = chai.expect;
+if (!expect) var expect = chai.expect;
 
 describe('Flatten function', function () {
     it('should properly deference a complex object including arrays', function () {
