+++ test/webpack.config.js
@@ -2,8 +2,8 @@ const webpack = require('webpack');
 const path = require('path');
 
 const core_config = {
-    mode: 'production',
-    devtool: 'none',
+    mode: 'development',
+    //devtool: 'none',
     entry: {
         bundle: './test/test.js',
     },
